refactor(web-ts-results): improve adaptability
